main/main/requirements.txt
django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-extra-fields==2.0.2
django-rest-auth==0.9.5
djangorestframework==3.10.3
pillow==6.2.1